Created abstraction for bubble sort view animations & moved helper function to commons file
